Payroll-Managenment/requirement.txt
Package         Version
--------------- -------
mysql-connector 2.2.9
tabulate        0.9.0
